AR-814: Refactored Template entity to be under tenant scope
